requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.8

# PDF generation (for payslips)
weasyprint==60.2
pdfkit==1.0.0

# AI and NLP (for production, replace with actual AI service)
openai==1.3.7
anthropic==0.7.8

# Security and encryption
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3 